+++ src/align/bounded/cloud_search_backward.rs
@@ -106,7 +106,7 @@ pub fn cloud_search_backward(
     cloud_matrix: &mut CloudMatrixLinear,
     params: &CloudSearchParams,
     bounds: &mut CloudBoundGroup,
-) -> anyhow::Result<()> {
+) {
     // the highest score we've seen overall
     let mut overall_max_score = -f32::INFINITY;
 
@@ -236,6 +236,4 @@ pub fn cloud_search_backward(
             }
         }
     }
-
-    Ok(())
 }