+++ src/align/bounded/structs/seed.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct Seed {
     pub target_name: String,
     pub target_start: usize,